move models to Entities folder
